Move ClearTextField out of Autofill
